Added archive capability

+++ lib/log_squirrel.rb
@@ -5,6 +5,7 @@ require 'yaml'
 require 'fog'
 require 'cfoundry'
 require 'tempfile'
+require 'zip/zip'
 
 class LogSquirrel
     
@@ -18,7 +19,7 @@ class LogSquirrel
 
   def upload_paths(job)
 
-    upload_time = Time.now.to_s
+    upload_time = Time.now.strftime '%Y%m%d-%H%M'
 
     storage = Fog::Storage.new @config[:fog_storage]
 
@@ -28,7 +29,7 @@ class LogSquirrel
     c.login creds
 
     app_name = job['application']
-    app = c.app_by_name app_name
+    app = c.app_by_name app_name 
 
     return if app.nil?
 
@@ -36,7 +37,11 @@ class LogSquirrel
 
     return if directory.nil?
 
-    backup_folder = directory.files.create :key => "#{app_name}-#{upload_time}/"
+    archive = job['archive']
+
+    backup_folder = directory.files.create :key => "#{app_name}-#{upload_time}/" if not archive
+
+    files_to_zip = {}
 
     job['paths'].each do |file_path|
 
@@ -50,18 +55,46 @@ class LogSquirrel
         tempfile.write(content)
         tempfile.close
 
-        # upload that resume
-        directory.files.create(
-          :key    => "#{backup_folder.key}#{local_path}",
-          :body   => File.open(tempfile.path),
-          :public => true
-        )
+        if archive
+          files_to_zip[local_path] = tempfile
+        else
 
-        tempfile.unlink
+          directory.files.create(
+            :key    => "#{backup_folder.key}#{local_path}",
+            :body   => File.open(tempfile.path),
+            :public => true
+          ) if not archive
+
+          tempfile.unlink
+        end
 
       rescue CFoundry::NotFound
         puts "404!"
       end
+
+    end
+
+    if not files_to_zip.empty?
+
+      temp_zip_file = Tempfile.new("#{app_name}.zip")
+      temp_folder = temp_zip_file.path.match(/(.+)\/([^\/]+)$/)[1]
+
+      temp_zip_path = "#{temp_folder}/#{app_name}-#{upload_time}.zip"
+
+      Zip::ZipFile.open(temp_zip_path, Zip::ZipFile::CREATE) do |zipfile|
+        files_to_zip.each do |local, temp_file|
+          zipfile.add(local, temp_file.path)
+        end
+      end
+      
+      directory.files.create(
+        :key    => "#{app_name}-#{upload_time}.zip",
+        :body   => File.open(temp_zip_path),
+        :public => true
+      )
+
+      files_to_zip.each { |path, file| file.unlink }
+      temp_zip_file.unlink
     end
 
   end
